requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
fastapi==0.110.0
Flask==3.0.2
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==3.0.2
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.36.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.27.1
Werkzeug==3.0.1
PyJWT==2.8.0
python-telegram-bot==20.8
